Add basic Jewel/Topaz chip support Fix Mifare Ultralight/C detection

+++ plugins/pluginsreaderproviders/pcsc/readers/id3readerunit.cpp
@@ -208,7 +208,7 @@ std::shared_ptr<Chip> ID3ReaderUnit::selectCard(std::shared_ptr<Chip> c)
     if (itr != chips_.end())
     {
         auto idx = std::distance(chips_.begin(), itr);
-        selectCard(idx);
+        selectCard(static_cast<uint8_t>(idx));
         return c;
     }
     return nullptr;
